Use double

+++ src/openflow/controllerApps/FiDeController.cc
@@ -79,8 +79,8 @@ void FiDeController::initialize(int stage){
             tunnels.push_back(tunnel);
 
             auto result = adjustment(tunnels);
-            float minLatency = result[result.size() - 1].minDelay;
-            float maxJitter = result[result.size() - 1].maxDelay - result[result.size() - 1].minDelay;
+            double minLatency = result[result.size() - 1].minDelay;
+            double maxJitter = result[result.size() - 1].maxDelay - result[result.size() - 1].minDelay;
             log("Result", INFO);
             log("minLatency: " + std::to_string(minLatency), INFO);
             log("maxJitter: " + std::to_string(maxJitter) , INFO);
